Improve keyring handling
In fact our little trick with writing garbage into the keyring and
setting the timeout afterwards doesn't seem to work. I noticed rather
late because I'm running this change with a higher timeout locally.

This is because the `key_update` in Linux resets `expiry` in this
case[1].

Originally I also tried to write into `@us` because that's shared
between `systemctl --user` and my Sway session, but it seems as if every
key created in there has the systemd user manager as possessor which
means that I can't do anything from the Sway session. This doesn't seem
to be changeable because newly created keys always have 0x3f010000 as
mask[2].

[1] https://github.com/torvalds/linux/blob/v6.8/security/keys/key.c#L1083
[2] https://github.com/torvalds/linux/blob/v6.8/security/keys/key.c#L905-L915